requirements.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
enum34==1.1.6
futures==3.1.1
gspread==0.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
ipaddress==1.0.18
oauth2client==1.5.2
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
pyOpenSSL==17.0.0
requests==2.18.1
requests-futures==0.9.7
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.10.0
tweepy==3.5.0
urllib3==1.21.1
